refactoring 05.30
