+++ router/api/upload.js
@@ -4,9 +4,9 @@ const multer = require("multer");
 const { v4: uuidv4 } = require('uuid');
 const path = require("path");
 
-const sendQuery = require("../feature/db");
-const feature = require("../feature/check");
-const requirement = require("../feature/requirement");
+const sendQuery = require("../../feature/db");
+const feature = require("../../feature/check");
+const requirement = require("../../feature/requirement");
 
 const upload = multer({
     storage: multer.diskStorage({
@@ -38,7 +38,7 @@ function checkFileType(file, cb){
   }
 
 
-router.post("/upload", async (req, res) => {
+router.post("/api/upload", async (req, res) => {
     const data = await requirement.getRequireData(req.session);
     if(!(await feature.checkAuth(req, res, data))) return;
