+++ static/js/api/delete.js
@@ -1,7 +1,7 @@
 $(".btn-delete-post").click(() => {
     if(confirm("해당 게시글을 삭제 하겠습니까?")){
         const post_idx = $("input[name='post-idx']").val();
-        deleteData(`/post/${post_idx}`);
+        deleteData(`/api/post/${post_idx}`);
     }
 })
