+++ views/fragment/header.ejs
@@ -14,7 +14,7 @@
             <% } else { %>
                 <% if(user_auth == "guest") { %><button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal-1">Auth</button> <% } %>
                 <!-- <a href="/profile"><button type="button" class="btn btn-secondary"><i class="far fa-address-card fa-lg"></i> Profile</button></a> -->
-                <button type="button" class="btn btn-danger feed <% if(feed == '1') { %> active <% } %>"><i class="fas fa-rss-square fa-lg"></i></button>
+                <button type="button" class="btn btn-danger feed <% if(require.feed == '1') { %> active <% } %>"><i class="fas fa-rss-square fa-lg"></i></button>
                 <a href="/write"><button type="button" class="btn btn-secondary">Write</button></a>
                 <a href="/logout"><button type="button" class="btn btn-secondary">Logout</button></a>
