+++ views/index.ejs
@@ -30,9 +30,9 @@
         </div>
         <%- include("./fragment/footer.ejs") -%>
 
-        <script src="/js/darkmode.js"></script>
-        <script src="/js/auth.js"></script>
-        <script src="/js/feed.js"></script>
+        <script src="/js/api/darkmode.js"></script>
+        <script src="/js/api/auth.js"></script>
+        <script src="/js/api/feed.js"></script>
         <script src="/js/HTML.js"></script>
         <script src="/js/api/index.js"></script>
     </body>
