+++ views/profile.ejs
@@ -50,6 +50,6 @@
                 This project is <a href="https://github.com/CASPER-REPSAC/project-manager">open source.</a>
             </div>
         </div>
-        <script src="/js/darkmode.js"></script>
+        <script src="/js/api/darkmode.js"></script>
     </body>
 </html>
