+++ views/write.ejs
@@ -132,11 +132,8 @@
         <%- include("./fragment/footer.ejs") -%>
 
         <script src="/js/HTML.js"></script>
-        <script src="/js/feature/writeForm.js"></script>
-        <script src="/js/darkmode.js"></script>
+        <script src="/js/api/darkmode.js"></script>
         <script src="/js/write.js"></script>
-        <script src="/js/feature/writeSubmit.js"></script>
-        <script src="/js/feature/uploadFile.js"></script>
         <script src="/js/feature/pdfToImg.js"></script>
         <script src="/js/uploadEvent.js"></script>
         <script src="/js/previewPDF.js"></script>
